Enum ruma_client_api::error::ErrorKind

#[non_exhaustive]
pub enum ErrorKind {

    Forbidden,
    UnknownToken {
        soft_logout: bool,
    },
    MissingToken,
    BadJson,
    NotJson,
    NotFound,
    LimitExceeded {
        retry_after_ms: Option<Duration>,
    },
    Unknown,
    Unrecognized,
    Unauthorized,
    UserDeactivated,
    UserInUse,
    InvalidUsername,
    RoomInUse,
    InvalidRoomState,
    ThreepidInUse,
    ThreepidNotFound,
    ThreepidAuthFailed,
    ThreepidDenied,
    ServerNotTrusted,
    UnsupportedRoomVersion,
    IncompatibleRoomVersion {
        room_version: RoomVersionId,
    },
    BadState,
    GuestAccessForbidden,
    CaptchaNeeded,
    CaptchaInvalid,
    MissingParam,
    InvalidParam,
    TooLarge,
    Exclusive,
    ResourceLimitExceeded {
        admin_contact: String,
    },
    CannotLeaveServerNoticeRoom,
    WeakPassword,
    UnableToAuthorizeJoin,
    UnableToGrantJoin,
    BadAlias,
    DuplicateAnnotation,
    NotYetUploaded,
    CannotOverwriteMedia,
    UnknownPos,
    UrlNotSet,
    BadStatus {
        status: Option<StatusCode>,
        body: Option<String>,
    },
    ConnectionFailed,
    ConnectionTimeout,
    WrongRoomKeysVersion {
        current_version: Option<String>,
    },

}

An enum for the error kind.

Items may contain additional information.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Forbidden

M_FORBIDDEN

UnknownToken

Fields

soft_logout: bool

If this is true, the client can acquire a new access token by specifying the device ID it is already using to the login API.

For more information, see the spec.

M_UNKNOWN_TOKEN

MissingToken

M_MISSING_TOKEN

BadJson

M_BAD_JSON

NotJson

M_NOT_JSON

NotFound

M_NOT_FOUND

LimitExceeded

Fields

retry_after_ms: Option<Duration>

How long a client should wait in milliseconds before they can try again.

M_LIMIT_EXCEEDED

Unknown

M_UNKNOWN

Unrecognized

M_UNRECOGNIZED

Unauthorized

M_UNAUTHORIZED

UserDeactivated

M_USER_DEACTIVATED

UserInUse

M_USER_IN_USE

InvalidUsername

M_INVALID_USERNAME

RoomInUse

M_ROOM_IN_USE

InvalidRoomState

M_INVALID_ROOM_STATE

ThreepidInUse

M_THREEPID_IN_USE

ThreepidNotFound

M_THREEPID_NOT_FOUND

ThreepidAuthFailed

M_THREEPID_AUTH_FAILED

ThreepidDenied

M_THREEPID_DENIED

ServerNotTrusted

M_SERVER_NOT_TRUSTED

UnsupportedRoomVersion

M_UNSUPPORTED_ROOM_VERSION

IncompatibleRoomVersion

Fields

room_version: RoomVersionId

The room’s version.

M_INCOMPATIBLE_ROOM_VERSION

BadState

M_BAD_STATE

GuestAccessForbidden

M_GUEST_ACCESS_FORBIDDEN

CaptchaNeeded

M_CAPTCHA_NEEDED

CaptchaInvalid

M_CAPTCHA_INVALID

MissingParam

M_MISSING_PARAM

InvalidParam

M_INVALID_PARAM

TooLarge

M_TOO_LARGE

Exclusive

M_EXCLUSIVE

ResourceLimitExceeded

Fields

admin_contact: String

A URI giving a contact method for the server administrator.

M_RESOURCE_LIMIT_EXCEEDED

CannotLeaveServerNoticeRoom

M_CANNOT_LEAVE_SERVER_NOTICE_ROOM

WeakPassword

M_WEAK_PASSWORD

UnableToAuthorizeJoin

M_UNABLE_TO_AUTHORISE_JOIN

UnableToGrantJoin

M_UNABLE_TO_GRANT_JOIN

BadAlias

M_BAD_ALIAS

DuplicateAnnotation

M_DUPLICATE_ANNOTATION

NotYetUploaded

M_NOT_YET_UPLOADED

CannotOverwriteMedia

M_CANNOT_OVERWRITE_MEDIA

UnknownPos

M_UNKNOWN_POS for sliding sync

UrlNotSet

M_URL_NOT_SET

BadStatus

Fields

status: Option<StatusCode>

The HTTP status code of the response.

body: Option<String>

The body of the response.

M_BAD_STATUS

ConnectionFailed

M_CONNECTION_FAILED

ConnectionTimeout

M_CONNECTION_TIMEOUT

WrongRoomKeysVersion

Fields

current_version: Option<String>

The currently active backup version.

M_WRONG_ROOM_KEYS_VERSION