Module ruma_client_api::push

Endpoints for push notifications.

Modules

delete_pushrule

DELETE /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}

get_notifications

GET /_matrix/client/*/notifications

get_pushers

GET /_matrix/client/*/pushers

get_pushrule

GET /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}

get_pushrule_actions

GET /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}/actions

get_pushrule_enabled

GET /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}/enabled

get_pushrules_all

GET /_matrix/client/*/pushrules/

get_pushrules_global_scope

GET /_matrix/client/*/pushrules/global/

set_pusher

POST /_matrix/client/*/pushers/set

set_pushrule

PUT /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}

set_pushrule_actions

PUT /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}/actions

set_pushrule_enabled

PUT /_matrix/client/*/pushrules/{scope}/{kind}/{ruleId}/enabled

Structs

EmailPusherData

Information for an email pusher.

MissingConditionsError

An error that happens when PushRule cannot be converted into ConditionalPushRule

MissingPatternError

An error that happens when PushRule cannot be converted into PatternedPushRule

PushRule

Like SimplePushRule, but may represent any kind of push rule thanks to pattern and conditions being optional.

Pusher

Defines a pusher.

PusherIds

Strings to uniquely identify a Pusher.

PusherInit

Initial set of fields of Pusher.

Enums

PusherKind

Which kind a pusher is, and the information for that kind.

RuleKind

The kinds of push rules that are available.

RuleScope

The scope of a push rule.