Module ruma_client_api::r0::uiaa

Module for User-Interactive Authentication API types.

Modules

authorize_fallback

GET /_matrix/client/r0/auth/{auth_type}/fallback/web?session={session_id}

Structs

AuthFlow

Description of steps required to authenticate via the User-Interactive Authentication API.

Dummy

Data for dummy UIAA flow.

EmailIdentity

Data for Email-based UIAA flow.

FallbackAcknowledgement

Data for UIAA fallback acknowledgement.

IncomingDummy

‘Incoming’ variant of Dummy.

IncomingEmailIdentity

‘Incoming’ variant of EmailIdentity.

IncomingFallbackAcknowledgement

‘Incoming’ variant of FallbackAcknowledgement.

IncomingMsisdn

‘Incoming’ variant of Msisdn.

IncomingOAuth2

‘Incoming’ variant of OAuth2.

IncomingPassword

‘Incoming’ variant of Password.

IncomingReCaptcha

‘Incoming’ variant of ReCaptcha.

IncomingThirdpartyIdCredentials

‘Incoming’ variant of ThirdpartyIdCredentials.

IncomingToken

‘Incoming’ variant of Token.

Msisdn

Data for phone number-based UIAA flow.

OAuth2

Data for OAuth2-based UIAA flow.

Password

Data for password-based UIAA flow.

ReCaptcha

Data for ReCaptcha UIAA flow.

ThirdpartyIdCredentials

Credentials for thirdparty authentification (e.g. email / phone number).

Token

Data for token-based UIAA flow.

UiaaInfo

Information about available authentication flows and status for User-Interactive Authenticiation API.

Enums

AuthData

Additional authentication information for the user-interactive authentication API.

IncomingAuthData

‘Incoming’ variant of AuthData.

IncomingUserIdentifier

‘Incoming’ variant of UserIdentifier.

UiaaResponse

Contains either a User-Interactive Authentication API response body or a Matrix error.

UserIdentifier

Identification information for the user.