Module ruma_client_api::sync::sync_events::v3

/v3/ (spec)

Structs

DeviceLists

Information on E2E device updates.

Ephemeral

Ephemeral events not recorded in the timeline or state of the room.

GlobalAccountData

The global private data created by this user.

IncomingRequest

‘Incoming’ variant of Request.

InviteState

The state of a room that the user has been invited to.

InvitedRoom

Updates to the rooms that the user has been invited to.

JoinedRoom

Updates to joined rooms.

KnockState

A mapping from a key events to a list of StrippedStateEvent.

KnockedRoom

Updates to knocked rooms.

LeftRoom

Historical updates to left rooms.

Presence

Updates to the presence status of other users.

Request

Data for a request to the sync API endpoint.

Response

Data in the response from the sync API endpoint.

RoomAccountData

The private data that this user has attached to this room.

RoomSummary

Information about room for rendering to clients.

Rooms

Updates to rooms.

State

State events in the room.

Timeline

Events in the room.

ToDevice

Messages sent directly between devices.

UnreadNotificationsCount

Unread notifications count.

Enums

Filter

A filter represented either as its full JSON definition or the ID of a saved filter.

IncomingFilter

‘Incoming’ variant of Filter.

Constants

METADATA

Metadata for the sync API endpoint.

Type Definitions

IncomingResponse

Convenience type alias for Response, for consistency with other IncomingResponse types.