Module ruma_client_api::uiaa

Module for User-Interactive Authentication API types.

Modules

get_uiaa_fallback_page

GET /_matrix/client/*/auth/{auth_type}/fallback/web?session={session_id}

Structs

AuthFlow

Description of steps required to authenticate via the User-Interactive Authentication API.

Dummy

Data for dummy UIAA flow.

EmailIdentity

Data for Email-based UIAA flow.

FallbackAcknowledgement

Data for UIAA fallback acknowledgement.

IncomingDummy

‘Incoming’ variant of Dummy.

IncomingEmailIdentity

‘Incoming’ variant of EmailIdentity.

IncomingFallbackAcknowledgement

‘Incoming’ variant of FallbackAcknowledgement.

IncomingMsisdn

‘Incoming’ variant of Msisdn.

IncomingPassword

‘Incoming’ variant of Password.

IncomingReCaptcha

‘Incoming’ variant of ReCaptcha.

IncomingRegistrationToken

‘Incoming’ variant of RegistrationToken.

Msisdn

Data for phone number-based UIAA flow.

Password

Data for password-based UIAA flow.

ReCaptcha

Data for ReCaptcha UIAA flow.

RegistrationToken

Data for registration token-based UIAA flow.

ThirdpartyIdCredentials

Credentials for third-party authentication (e.g. email / phone number).

UiaaInfo

Information about available authentication flows and status for User-Interactive Authenticiation API.

Enums

AuthData

Information for one authentication stage.

AuthType

The type of an authentication stage.

IncomingAuthData

‘Incoming’ variant of AuthData.

IncomingUserIdentifier

‘Incoming’ variant of UserIdentifier.

UiaaResponse

Contains either a User-Interactive Authentication API response body or a Matrix error.

UserIdentifier

Identification information for the user.